Rename DealsChart component to DealsAreaChart
